feat: add events and refactor
