+++ packages/node/src/index.ts
@@ -4,3 +4,4 @@ import { PushProviderIdEnum } from '@novu/shared';
 export { ChatProviderIdEnum, PushProviderIdEnum };
 export * from './lib/novu';
 export * from './lib/subscribers/subscriber.interface';
+export * from './lib/events/events.interface';
