+++ packages/node/src/lib/events/events.spec.ts
@@ -1,0 +1,18 @@
+import { Novu } from '../novu';
+import axios from 'axios';
+
+const mockConfig = {
+  apiKey: '1234',
+};
+
+jest.mock('axios');
+
+describe('test use of novus node package - Events', () => {
+  const mockedAxios = axios as jest.Mocked<typeof axios>;
+  let novu: Novu;
+
+  beforeEach(() => {
+    mockedAxios.create.mockReturnThis();
+    novu = new Novu(mockConfig.apiKey);
+  });
+});
